requirements.txt
aiocontextvars==0.2.2
argcomplete==3.6.2
brotli==1.1.0
brotlicffi==1.1.0.0
ConfigParser==7.2.0
contextlib2==21.6.0
cryptography==45.0.3
docutils==0.21.2
exceptiongroup==1.3.0
h2==4.2.0
HTMLParser==0.0.2
importlib_metadata==8.0.0
ipykernel==6.29.5
ipython==8.12.3
ipywidgets==8.1.7
Jinja2==3.1.6
jnius==1.1.0
js==1.0
keyring==25.6.0
lockfile==0.12.2
mock==5.2.0
numpy==2.2.6
objgraph==3.6.2
pexpect==4.9.0
Pillow==11.2.1
protobuf==6.31.1
psutil==7.0.0
PyInstaller==6.13.0
pyodide==0.0.2
pyOpenSSL==25.1.0
pytest_metadata==3.1.1
pytest_xdist==3.7.0
PyYAML==5.3.1
PyYAML==6.0.2
railroad==0.5.0
simplejson==3.20.1
Sphinx==8.2.3
thread==2.0.5
tornado==6.5.1
truststore==0.10.1
Unidecode==1.4.0
urllib3_secure_extra==0.1.0
xmlrpclib==1.0.1
zope==5.13
zstandard==0.23.0
